pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cortex"
version = "0.1.0"
description = "Agentic local LLM voice assistant for Raspberry Pi 5 + NPU"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    # Core
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "websockets>=12.0",
    # IPC
    "pyzmq>=25.0",
    # Database
    "aiosqlite>=0.19.0",
    # Audio
    "numpy>=1.26.0",
    "sounddevice>=0.4.6",
    # HTTP client (for LLM API subprocess communication)
    "httpx>=0.25.0",
    # CLI
    "click>=8.1.0",
    # Web UI
    "jinja2>=3.1.0",
    "python-multipart>=0.0.6",
    "bcrypt>=4.0.0",
    "itsdangerous>=2.1.0",
    # Utilities
    "pyyaml>=6.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "structlog>=23.2.0",
    "rich>=13.7.0",
    # External services (Phase 3b)
    "caldav>=1.3.0",
    # MCP protocol (Phase 3b)
    "mcp>=1.0.0",
    # Embedding tokenization (Phase 4)
    "tokenizers>=0.15.0",
    # Document ingestion (Phase 4)
    "pypdf>=3.17.0",
    "beautifulsoup4>=4.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "pre-commit>=3.6.0",
    "types-PyYAML>=6.0",
]
pi = [
    # Pi-specific deps installed only on target hardware
    "uvloop>=0.19.0",
    "onnxruntime>=1.16.0",
]

[project.scripts]
cortex = "cortex.cli:main"
cortex-npu = "cortex.hal.npu.main:main"
cortex-audio = "cortex.hal.audio.main:main"
cortex-display = "cortex.hal.display.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true

[[tool.mypy.overrides]]
module = [
    "sounddevice.*", "zmq.*", "axengine.*", "bcrypt.*", "caldav.*", "mcp.*",
    "onnxruntime.*", "tokenizers.*", "pypdf.*", "bs4.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "hardware: requires Pi hardware (deselect with '-m not hardware')",
]
